feat/approval: #109 결재 생성 구현 중

+++ perflow-backend/src/main/java/com/touchdown/perflowbackend/approval/command/application/dto/ApproveLineDTO.java
@@ -21,7 +21,7 @@ public class ApproveLineDTO {
     private final ApproveType approveType;
 
     @JsonProperty("order")
-    private final Integer approveLineOrder;
+    private final Long approveLineOrder;
 
     @JsonProperty("pllGroupId")
     private final Long pllGroupId;
